Renames the directory and test class name to follow the project pattern Creates unit test for the function of removing item from the list Merge develop branch modifications Fixes bug in date of birth formatting Implements function to delete pet from API Changes type of the ID variable from Long to String Fixes some bugs in the code Fixes bug when saving a pet in the database Adds pull refresh to the list of pets

+++ petJournal/app/src/main/java/com/soujunior/petjournal/ui/screens_app/screens_pets/petRaceAndSizeScreen/components/Screen.kt
@@ -54,8 +54,8 @@ fun Screen(idPetInformation: String?, navController: NavController) {
     val taskState by viewModel.taskState.collectAsState()
     val isTextFiledOthersVisible by viewModel.isTextFiledOthersVisible.collectAsState()
     if (idPetInformation != null) {
-        viewModel.getPetInformation(idPetInformation.toLong())
-        RaceSizeFormEvent.IdPetInformation(idPetInformation = idPetInformation.toLong())
+        viewModel.getPetInformation(idPetInformation)
+        RaceSizeFormEvent.IdPetInformation(idPetInformation = idPetInformation)
     }
     val scrollState = rememberLazyListState()
     val screenHeight = getScreenHeightInch()
@@ -99,7 +99,6 @@ fun Screen(idPetInformation: String?, navController: NavController) {
                                 Breadcrumb(index = 2)
                             }
                             item {
-
                                 Box(
                                     modifier = Modifier
                                         .fillMaxWidth()
@@ -114,7 +113,6 @@ fun Screen(idPetInformation: String?, navController: NavController) {
                                 }
 
                             }
-
 
                             item {
                                 DropDown(
